refactor: ReportStatus enum을 독립 파일로 분리하여 도메인 패키지로 이동
