+++ src/main/java/com/b4f2/pting/domain/ReportStatus.java
@@ -1,0 +1,7 @@
+package com.b4f2.pting.domain;
+
+public enum ReportStatus {
+    PENDING,   // 처리 전
+    RESOLVED,  // 처리 완료
+    REJECTED   // 기각
+}
