+++ src/main/java/com/b4f2/pting/dto/GameReportStatusUpdateRequest.java
@@ -1,6 +1,6 @@
 package com.b4f2.pting.dto;
 
-import com.b4f2.pting.domain.GameReport.ReportStatus;
+import com.b4f2.pting.domain.ReportStatus;
 
 public record GameReportStatusUpdateRequest(
     ReportStatus status
